+++ components/pages/discovery/components/Facets/Folder.tsx
@@ -18,20 +18,8 @@
  */
 
 import { css } from '@emotion/react';
-import { Icon, Input } from '@icgc-argo/uikit';
 import { PropsWithChildren, useEffect, useState } from 'react';
-
-const commonStyle = {
-  header: css({
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-
-    '> div, svg': {
-      marginLeft: 'auto',
-    },
-  }),
-};
+import { ArrowToggle, commonStyle } from './common';
 
 const facetFolderStyles = {
   container: css({ ':hover': { cursor: 'pointer' } }),
@@ -43,10 +31,6 @@ const facetFolderStyles = {
     },
   ]),
   content: (isOpen) => ({ display: isOpen ? 'block' : 'none' }),
-};
-
-const ArrowToggle = ({ isOpen }) => {
-  return <Icon name={isOpen ? 'chevron_up' : 'chevron_down'} fill="white" height="8px" />;
 };
 
 export const FacetFolder = ({
@@ -81,39 +65,4 @@ export const FacetFolder = ({
       <div css={facetFolderStyles.content(isOpen)}>{children}</div>
     </div>
   );
-};
-
-const filtersSearchBoxStyles = {
-  folder: css({
-    h2: { margin: 0, fontSize: '16px', fontWeight: 400 },
-  }),
-};
-
-export const FiltersSearchBox = ({ title, onClick, isExpanded }) => {
-  return (
-    <div css={css([css({ height: '100px', margin: '10px 8px 8px 8px' })])}>
-      <div css={css([commonStyle.header, filtersSearchBoxStyles.folder])}>
-        <h2>{title}</h2>
-        <div onClick={() => onClick()}>
-          Expand All <ArrowToggle isOpen={isExpanded} />
-        </div>
-      </div>
-
-      <Input
-        size="sm"
-        aria-label="search-for-files"
-        placeholder={'try searcing for a filter'}
-        preset="search"
-        value={'1'}
-        onChange={(e) => {
-          console.log(e.target.value);
-        }}
-        css={css({ marginTop: '12px' })}
-      />
-    </div>
-  );
-};
-
-export const FacetGroup = ({ children }: PropsWithChildren<{}>) => {
-  return <div>{children}</div>;
 };
